feat: productInfo컴포넌트 분리
